flatten out intenral and shared (1/2)
