+++ packages/core/src/http/types.ts
@@ -2,7 +2,8 @@ import type { TRPCError } from '../error/TRPCError';
 import type { ProcedureType } from '../procedure';
 import type { AnyRouter } from '../router';
 import type { TRPCResponse } from '../rpc';
-import type { Dict, inferRouterContext, inferRouterError } from '../types';
+import type { inferRouterContext, inferRouterError } from '../types';
+import type { Dict } from '../utilityTypes';
 
 export type HTTPHeaders = Dict<string[] | string>;
