+++ packages/core/src/internals/mergeRouters.ts
@@ -3,7 +3,7 @@ import type { AnyRouter, AnyRouterDef, Router, RouterDef } from '../router';
 import { createRouterFactory } from '../router';
 import type { CombinedDataTransformer } from '../transformer';
 import { defaultTransformer } from '../transformer';
-import { mergeWithoutOverrides } from './mergeWithoutOverrides';
+import { mergeWithoutOverrides } from '../utilityFunctions';
 
 /**
  * @internal
