--- packages/core/src/observable/internals/identity.ts
@@ -1,3 +1,0 @@
-export function identity<TType>(x: TType): TType {
-  return x;
-}
