--- packages/core/src/observable/operators/index.ts
@@ -1,3 +1,0 @@
-export * from './share';
-export * from './map';
-export * from './tap';
