+++ packages/core/src/shared/getErrorShape.ts
@@ -1,8 +1,8 @@
 import type { DefaultErrorShape } from '../error/formatter';
 import type { TRPCError } from '../error/TRPCError';
 import { getHTTPStatusCodeFromError } from '../http/getHTTPStatusCode';
-import type { AnyRootConfig } from '../internals/config';
 import type { ProcedureType } from '../procedure';
+import type { AnyRootConfig } from '../rootConfig';
 import { TRPC_ERROR_CODES_BY_KEY } from '../rpc';
 
 /**
