+++ packages/core/src/types.ts
@@ -75,86 +75,7 @@ export type ProtectedIntersection<TType, TWith> = keyof TType &
   : IntersectionError<string & keyof TType & keyof TWith>;
 
 /**
- * @public
- */
-export type Maybe<TType> = TType | null | undefined;
-
-/**
  * @internal
- * @see https://github.com/ianstormtaylor/superstruct/blob/7973400cd04d8ad92bbdc2b6f35acbfb3c934079/src/utils.ts#L323-L325
+ * Returns the raw input type of a procedure
  */
-export type Simplify<TType> = TType extends any[] | Date
-  ? TType
-  : { [K in keyof TType]: TType[K] };
-/**
- * @public
- */
-export type Dict<TType> = Record<string, TType | undefined>;
-
-/**
- * @public
- */
-export type MaybePromise<TType> = Promise<TType> | TType;
-
-/**
- * @internal
- *
- * Creates a "lower-priority" type inference.
- * https://github.com/microsoft/TypeScript/issues/14829#issuecomment-322267089
- */
-export type InferLast<TType> = TType & {
-  [KeyType in keyof TType]: TType[KeyType];
-};
-
-export type FilterKeys<TObj extends object, TFilter> = {
-  [TKey in keyof TObj]: TObj[TKey] extends TFilter ? TKey : never;
-}[keyof TObj];
-
-/**
- * @internal
- */
-export type Filter<TObj extends object, TFilter> = Pick<
-  TObj,
-  FilterKeys<TObj, TFilter>
->;
-
-/**
- * Unwrap return type if the type is a function (sync or async), else use the type as is
- * @internal
- */
-export type Unwrap<TType> = TType extends (...args: any[]) => infer R
-  ? Awaited<R>
-  : TType;
-
-/**
- * Makes the object recursively optional
- * @internal
- */
-export type DeepPartial<TObject> = TObject extends object
-  ? {
-      [P in keyof TObject]?: DeepPartial<TObject[P]>;
-    }
-  : TObject;
-
-/**
- * Omits the key without removing a potential union
- * @internal
- */
-export type DistributiveOmit<TObj, TKey extends keyof any> = TObj extends any
-  ? Omit<TObj, TKey>
-  : never;
-
-/*
- * See https://github.com/microsoft/TypeScript/issues/41966#issuecomment-758187996
- * Fixes issues with iterating over keys of objects with index signatures.
- * Without this, iterations over keys of objects with index signatures will lose
- * type information about the keys and only the index signature will remain.
- * @internal
- */
-export type WithoutIndexSignature<TObj> = {
-  [K in keyof TObj as string extends K
-    ? never
-    : number extends K
-    ? never
-    : K]: TObj[K];
-};
+export type GetRawInputFn = () => Promise<unknown>;
